pyproject.toml
[project]
name = "gerby-web"
version = "0.0.1"
description = "."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "gerby @ git+https://github.com/litProgLaTeX/lpil-gerby-website.git",
    "setuptools>=69.1.1",
    "waitress>=3.0.0",
    "plasTeX @ git+https://github.com/diSimplex/plastex.git@merged",
    "lpilGerbyConfig @ git+https://github.com/litProgLaTeX/lpil-gerby-configuration.git",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
gerbyRunner   = "gerbyRunner.runner:cli"
gerbyCompiler = "gerbyRunner.compiler:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
post_install = {	call = "gerbyRunner.postInstall:patches" }

